Extract filter chip tray and create context
